Modify the Build & Publish GitHub Action to use Dockerfile from sneezymud-docker repo
Currently, the Build & Publish action that creates and uploads new Docker images to Docker Hub when changes are merged to master points to a Dockerfile that's part of this repo. This Dockerfile is an identical copy to one in the sneezymud-docker repo in the docker folder.

This PR modifies the Build & Publish action to use the Dockerfile from the sneezymud-docker repo instead, and deletes the Dockerfile in this repo, making the version in sneezymud-docker the single source of truth for building images for the sneezy container.

As things are currently set up, we're forced to make an identical commit to sneezymud-docker any time we modify the Dockerfile in this repo if we want to make sure they're kept in sync.