Modify Build and Test Github workflow to ensure it's always run
The master branch's settings require every PR submitted to have the 'Build and Test' GitHub action defined in .github/workflows/test.yml to run and pass before the PR can be merged.

Recently I added some path filters to test.yml that only trigger the workflow when relevant files are modified in a PR, to avoid having to wait for compilation to finish just to submit modifications to things like static zonefiles.

However, this is now preventing this workflow from executing at all for PRs like that, which is blocking them from being mergeable. This change should hopefully solve this issue.